python/python-kafka-rabbitmq-programming-examples/requirements.txt
kafka-python
websockets
pika